refactor: split the factory file into 2 files
- standard-config.ts
- workspace-config.ts
