+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const NGRX_PACKAGES = ['@ngrx/store', '@ngrx/effects', '@ngrx/signals', '@ngrx/operators'] as const;
