+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './configs';
-export * from './factory';
 export { GLOB_HTML, GLOB_JS, GLOB_SRC, GLOB_TESTS, GLOB_TS } from './globs';
+export * from './standard-config';
 export type * from './types';
 export * from './utils';
+export * from './workspace-config';
